refactor(ai): reorganize commands module into logical submodules
Split the monolithic commands.rs file (905 lines) into organized submodules,
improving code maintainability and reducing cognitive load. Each module now
focuses on a specific domain of functionality.

New module structure:
- commands/mod.rs: AiState, shared helpers, re-exports
- commands/core.rs: Core AI agent initialization and execution
- commands/config.rs: LLM provider setup and configuration
- commands/session.rs: Session/conversation management
- commands/hitl.rs: Human-in-the-loop approval system
- commands/policy.rs: Tool policy and auto-execution controls
- commands/context.rs: Context window and token management
- commands/loop_detection.rs: Loop detection and protection

Benefits:
- Better organization: Related commands grouped logically
- Easier navigation: Developers can quickly find commands by domain
- Reduced complexity: Each file is focused and easier to understand
- Improved maintainability: Changes to one domain don't affect others
- All 153 tests pass: No functionality lost in the refactoring

+++ src-tauri/src/ai/agentic_loop.rs
@@ -45,7 +45,7 @@ pub const MAX_TOOL_ITERATIONS: usize = 100;
 pub const APPROVAL_TIMEOUT_SECS: u64 = 300;
 
 /// Maximum tokens for a single completion request
-pub const MAX_COMPLETION_TOKENS: i32 = 10_000;
+pub const MAX_COMPLETION_TOKENS: u32 = 10_000;
 
 /// Context for the agentic loop execution.
 pub struct AgenticLoopContext<'a> {
@@ -484,7 +484,7 @@ pub async fn run_agentic_loop(
             documents: vec![],
             tools: tools.clone(),
             temperature: Some(0.5),
-            max_tokens: Some(MAX_COMPLETION_TOKENS),
+            max_tokens: Some(MAX_COMPLETION_TOKENS as u64),
             tool_choice: None,
             additional_params: None,
         };
